Database.toSchema doesn’t belong to the public API
We have Database.load/save, and that should be enough.

For more advanced situations, there’s always scala.meta.internal.semantic,
but that’s not something that I’d like to guarantee to support in the future.
